# src/main/java/com/example/auth/adapter/in/web/dto/LoginRequest.java
package com.example.auth.adapter.in.web.dto;

/**
 * Request DTO for user login.
 */
public class LoginRequest {
    public String email;
    public String password;
    
    public LoginRequest() {}
    
    public LoginRequest(String email, String password) {
        this.email = email;
        this.password = password;
    }
    
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
    
    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }
}