# src/main/java/com/example/auth/adapter/in/web/dto/SignupRequest.java
package com.example.auth.adapter.in.web.dto;

/**
 * Request DTO for user signup.
 */
public class SignupRequest {
    public String email;
    public String password;
    public String fullName;
    
    public SignupRequest() {}
    
    public SignupRequest(String email, String password, String fullName) {
        this.email = email;
        this.password = password;
        this.fullName = fullName;
    }
    
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
    
    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }
    
    public String getFullName() { return fullName; }
    public void setFullName(String fullName) { this.fullName = fullName; }
}